# perf: use faster form of copying dates, across the board improvement