Reference: http://jsperf.com/date-copy-constructor/3

Didn't expect this to make so much of a difference, but here we are.  I presume `Moment` always gets a `config` object where `_d` is a Date; certainly seems to be the case and no tests failed.  Results of `grunt test update-index benchmark`:

-----
Before:
```
Running benchmark clone [benchmarks/clone.js]...
>> clone x 3,698,440 ops/sec ±0.77% (85 runs sampled)
Running benchmark fromDate [benchmarks/fromDate.js]...
>> fromDate x 1,709,634 ops/sec ±1.05% (79 runs sampled)
Running benchmark fromDateUtc [benchmarks/fromDateUtc.js]...
>> fromDateUtc x 1,623,291 ops/sec ±0.95% (90 runs sampled)
```

-----
After:
```
Running benchmark clone [benchmarks/clone.js]...
>> clone x 4,919,547 ops/sec ±0.69% (87 runs sampled)
Running benchmark fromDate [benchmarks/fromDate.js]...
>> fromDate x 2,232,456 ops/sec ±0.54% (92 runs sampled)
Running benchmark fromDateUtc [benchmarks/fromDateUtc.js]...
>> fromDateUtc x 2,151,327 ops/sec ±1.01% (78 runs sampled)
```