@ichernev 2015-07-03
Good catch. How much of the speed improvement is from the removal of `console.log` :)

I'll merge it in next release.