@wyantb 2015-07-04
:) presumably none, since the after call shouldn't be included in timings
by Benchmark.  Still did a double take at your comment, though, hah.
On Jul 3, 2015 6:44 AM, "Iskren Ivov Chernev" <notifications@github.com>
wrote:

> Good catch. How much of the speed improvement is from the removal of
> console.log :)
>
> I'll merge it in next release.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/moment/moment/pull/2368#issuecomment-118313696>.
>