@ichernev 2015-07-13
Merged in b6b7b87337376cfdc0e40da7b74e4afb664c0185